Major updates to 2.0.0

+++ lib/widgets/audio_recorder_dialog.dart
@@ -1,8 +1,8 @@
+// ignore_for_file: depend_on_referenced_packages
 import 'dart:io';
-
 import 'package:file/local.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_audio_recorder/flutter_audio_recorder.dart';
+import 'package:flutter_audio_recorder3/flutter_audio_recorder3.dart';
 import 'package:material_themes_manager/material_themes_manager.dart';
 import 'package:material_themes_widgets/forms/form_fields.dart';
 import 'package:material_themes_widgets/fundamental/icons.dart';
@@ -11,21 +11,24 @@ import 'package:material_themes_widgets/utils/validators.dart';
 import 'package:path/path.dart' as path;
 import 'package:path_provider/path_provider.dart';
 import 'package:provider/provider.dart';
+import 'dart:developer' as developer;
 
 class AudioRecorderDialog extends StatefulWidget {
   final LocalFileSystem localFileSystem;
 
-  AudioRecorderDialog({localFileSystem}) : this.localFileSystem = localFileSystem ?? LocalFileSystem();
-  _AudioRecorderDialogState createState() => _AudioRecorderDialogState();
+  const AudioRecorderDialog({super.key, localFileSystem}) : localFileSystem = localFileSystem ?? const LocalFileSystem();
+
+  @override
+  AudioRecorderDialogState createState() => AudioRecorderDialogState();
 }
 
-class _AudioRecorderDialogState extends State<AudioRecorderDialog> {
+class AudioRecorderDialogState extends State<AudioRecorderDialog> {
 
   final _formKey = GlobalKey<FormState>();
-  FlutterAudioRecorder _recorder;
-  RecordingStatus _recordingStatus;
-  Directory _directory;
-  String recordingName;
+  late FlutterAudioRecorder3 _recorder;
+  RecordingStatus _recordingStatus = RecordingStatus.Unset;
+  Directory? _directory;
+  String? recordingName;
 
   @override
   void initState() {
@@ -38,7 +41,7 @@ class _AudioRecorderDialogState extends State<AudioRecorderDialog> {
     return AlertDialog(
       //contentPadding: EdgeInsets.all(0.0),
       content: Stack(
-        overflow: Overflow.visible,
+        clipBehavior: Clip.none,
         children: <Widget>[
           //context.watch<MaterialThemesManager>().getBackgroundGradient(BackgroundGradientType.PRIMARY),
           Positioned(
@@ -49,8 +52,8 @@ class _AudioRecorderDialogState extends State<AudioRecorderDialog> {
                 _quitWithoutSaving();
               },
               child: CircleAvatar(
-                child: ThemedIcon(Icons.close, type: ThemeGroupType.MOP),
                 backgroundColor: context.watch<MaterialThemesManager>().colorPalette().primary,
+                child: ThemedIcon(Icons.close, type: ThemeGroupType.MOP),
               ),
             ),
           ),
@@ -61,34 +64,36 @@ class _AudioRecorderDialogState extends State<AudioRecorderDialog> {
               children: <Widget>[
                 if (_recordingStatus == RecordingStatus.Recording) ... [
                   Padding(
-                      padding: EdgeInsets.all(8.0),
-                      child: ThemedIconButton(Icons.pause, onPressedCallback: () => setState(() => {
-                        _pause()
-                      }))
+                      padding: const EdgeInsets.all(8.0),
+                      child: ThemedIconButton(Icons.pause, onPressedCallback: () => setState(() => _pause()
+                      ))
                   )
                 ]
                 else ... [
                   Padding(
-                      padding: EdgeInsets.all(8.0),
+                      padding: const EdgeInsets.all(8.0),
                       child: ThemedIconButton(Icons.mic, onPressedCallback: () => _record())
                   )
                 ],
                 ThemedEditableLabelValue(
                   showLabel: false,
                   hintText: "Recording Name",
-                  validator:  (value) => nonEmptyValidator(value, "Name"),
-                  onStringChangedCallback: (value) => setState(() => { recordingName = value }),
+                  validator:  (value) => nonEmptyValidator(value, "Name", true),
+                  onStringChangedCallback: (value) => setState(() => recordingName = value ),
                 ),
                 if (_recordingStatus != RecordingStatus.Initialized) ... [
                   Padding(
                     padding: const EdgeInsets.all(8.0),
-                    child: RaisedButton(
-                      shape: StadiumBorder(),
-                      color: context.watch<MaterialThemesManager>().colorPalette().primary,
+                    child: ElevatedButton(
+                      style: ElevatedButton.styleFrom(
+                        padding: const EdgeInsets.symmetric(horizontal: 40.0, vertical: 20.0),
+                        backgroundColor: context.watch<MaterialThemesManager>().colorPalette().primary,
+                        shape: const StadiumBorder(),
+                      ),
                       child: ThemedTitle("Save", type: ThemeGroupType.MOP),
                       onPressed: () {
-                        if (_formKey.currentState.validate()) {
-                          _formKey.currentState.save();
+                        if (_formKey.currentState?.validate() == true) {
+                          _formKey.currentState?.save();
 
                           //There is a valid file name, so save the recording and close the dialog
                           _save();
@@ -111,26 +116,30 @@ class _AudioRecorderDialogState extends State<AudioRecorderDialog> {
 
   void _initAsync() async {
     _directory = await getApplicationDocumentsDirectory();
-    final tempPath = '${_directory.path}/tempAudioRecording';
 
-    //Delete the last temp recording if it exists - no reason to keep these around
-    try {
-      final file = File(tempPath);
-      await file.delete();
-    } catch (e) {
+    if (_directory != null) {
+      final tempPath = '${_directory?.path}/tempAudioRecording';
+
+      //Delete the last temp recording if it exists - no reason to keep these around
+      try {
+        final file = File(tempPath);
+        await file.delete();
+      } catch (e) {
+        developer.log(e.toString());
+      }
+
+      _recorder = FlutterAudioRecorder3(tempPath, audioFormat: AudioFormat.AAC); // or AudioFormat.WAV
+      await _recorder.initialized;
+      var recording = await _recorder.current(channel: 0);
+      setRecordingStatus(recording);
     }
-
-    _recorder = FlutterAudioRecorder(tempPath, audioFormat: AudioFormat.AAC); // or AudioFormat.WAV
-    await _recorder.initialized;
-    var recording = await _recorder.current(channel: 0);
-    setState(() { _recordingStatus = recording?.status; });
   }
 
   void _pause() async {
     await _recorder.pause();
 
     var recording = await _recorder.current(channel: 0);
-    setState(() { _recordingStatus = recording?.status; });
+    setRecordingStatus(recording);
   }
 
   void _record() async {
@@ -139,7 +148,7 @@ class _AudioRecorderDialogState extends State<AudioRecorderDialog> {
         : await _recorder.start();
 
     var recording = await _recorder.current(channel: 0);
-    setState(() { _recordingStatus = recording?.status; });
+    setRecordingStatus(recording);
   }
 
   void _quitWithoutSaving() async {
@@ -148,25 +157,37 @@ class _AudioRecorderDialogState extends State<AudioRecorderDialog> {
     }
 
     var recording = await _recorder.current(channel: 0);
-    setState(() { _recordingStatus = recording?.status; });
+    setRecordingStatus(recording);
 
-    Navigator.pop(context, "");
+    if (mounted) Navigator.pop(context, "");
   }
 
   void _save() async {
     var result = await _recorder.stop();
     var recording = await _recorder.current(channel: 0);
-    setState(() { _recordingStatus = recording?.status; });
+    setRecordingStatus(recording);
 
-    String ext = path.extension(result.path).substring(1);//Need to string dot from .extension
-    String dir = path.dirname(result.path);
-    String newPath = path.join(dir, '$recordingName.$ext');
+    String? resultPath = result?.path;
+    if (resultPath != null && recordingName != null) {
+      String ext = path.extension(resultPath).substring(1);//Need to string dot from .extension
+      String dir = path.dirname(resultPath);
+      String newPath = path.join(dir, '$recordingName.$ext');
 
-    print("NewPath:" + newPath + " ext:" + ext + " dir:" + dir);
+      developer.log("NewPath:$newPath ext:$ext dir:$dir");
 
-    File file = File(result.path);
-    await file.rename(newPath);
+      File file = File(resultPath);
+      await file.rename(newPath);
 
-    Navigator.pop(context, newPath);
+      if (mounted) Navigator.pop(context, newPath);
+    } else {
+      if (mounted) Navigator.pop(context);
+    }
+  }
+
+  void setRecordingStatus(Recording? recording) {
+    var recordingStatus = recording?.status;
+    if (recordingStatus != null) {
+      setState(() { _recordingStatus = recordingStatus; });
+    }
   }
 }
